Odds n Ends
Added dialog.notice, which makes messages on the fly without an entry
from a resource.
Added descriptions to control sets and implimented them in an easier way
to handle the controls button.
Added some SETMEs to g.ui for the about and controls dialogs.
Movied serialization and saving to g.save instead of g.ui.